python/flask-website/webapp/requirements.txt
Flask==2.0.1
mysql-connector==2.2.9
Flask-Bootstrap==3.3.7.1
